Moved Documents LandingPageComponent Login NewManagementPlan to App folder
