+++ submit-web/src/components/App/NewManagementPlan/PlanDetails.tsx
