--- submit-web/src/components/Shared/CustomRadioOptions.tsx
@@ -1,41 +1,0 @@
-import { FormControlLabel, Radio } from "@mui/material";
-import { BCDesignTokens } from "epic.theme";
-
-type Option = {
-  label: string;
-  value: string | number | boolean;
-};
-
-type CustomRadioOptionsProps = {
-  options: Option[];
-  error?: boolean;
-  disabled?: boolean;
-};
-
-export const CustomRadioOptions = ({
-  options,
-  error = false,
-  disabled = false,
-}: CustomRadioOptionsProps) => {
-  const sx = [
-    disabled && {
-      color: `${BCDesignTokens.typographyColorDisabled} !important`,
-    },
-    error && {
-      color: BCDesignTokens.surfaceColorPrimaryDangerButtonDefault,
-    },
-  ];
-
-  return (
-    <>
-      {options.map(({ label, value }) => (
-        <FormControlLabel
-          key={value.toString()}
-          value={value}
-          control={<Radio sx={sx} disabled={disabled} />}
-          label={label}
-        />
-      ))}
-    </>
-  );
-};
