--- submit-web/src/components/SubmissionItem/ConsultationRecord/ConsultationRecordProponentView/radioOptions.tsx
@@ -1,27 +1,0 @@
-import { RadioOptions } from "@/components/Shared/controlled/ControlledRadioGroup";
-
-export const YesOrNoOptions: RadioOptions[] = [
-  {
-    label: "Yes",
-    value: true,
-  },
-  {
-    label: "No",
-    value: false,
-  },
-];
-
-export const YesNoNotApplicableOptions: RadioOptions[] = [
-  {
-    label: "Yes",
-    value: false,
-  },
-  {
-    label: "No",
-    value: true,
-  },
-  {
-    label: "Not Applicable",
-    value: "",
-  },
-];
