+++ submit-web/src/routes/index.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute, Navigate } from "@tanstack/react-router";
 import { useAuth } from "react-oidc-context";
 import { PageLoader } from "@/components/Shared/PageLoader";
-import { LandingPageComponent } from "@/components/LandingPageComponent";
+import { LandingPageComponent } from "@/components/App/LandingPageComponent";
 
 export const Route = createFileRoute("/")({
   component: Index,
